+++ tests/test_parse.py
@@ -7,7 +7,7 @@ import unittest
 from pythainlp.parse import dependency_parsing
 
 
-class TestParsePackage(unittest.TestCase):
+class ParseTestCase(unittest.TestCase):
     def test_dependency_parsing(self):
         self.assertIsNotNone(dependency_parsing("ผมเป็นคนดี", engine="esupar"))
         self.assertIsNotNone(
